pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shipguard"
version = "0.2.0"
description = "Reusable security audit tool for any repository"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "DevOpsCelstn" }]
keywords = ["security", "audit", "cli", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pathspec>=0.11",
]

[project.urls]
Homepage = "https://github.com/celstnblacc/shipguard"
Repository = "https://github.com/celstnblacc/shipguard"
Changelog = "https://github.com/celstnblacc/shipguard/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/celstnblacc/shipguard/issues"
"Security Policy" = "https://github.com/celstnblacc/shipguard/blob/main/SECURITY.md"

[project.scripts]
shipguard = "shipguard.cli:app"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "hypothesis>=6.0"]

[project.entry-points."shipguard.rules"]
# Third-party rule packages register here

[tool.hatch.build.targets.wheel]
packages = ["src/shipguard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "performance: performance regression tests (usually gated by env var)",
    "mutation: mutation test harness tests (usually gated by env var)",
]
